add detail/parsing/lexer.hpp
